requirements.txt
aniso8601==8.0.0
Click==7.0
colorama==0.4.3
coverage==5.0.3
entrypoints==0.3
flake8==3.7.9
flake8-polyfill==1.0.2
Flask==1.1.1
Flask-RESTful==0.3.7
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.10.3
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
nose==1.3.7
pycodestyle==2.5.0
pycrypto==2.6.1
pyflakes==2.1.1
python-slugify==4.0.0
pytz==2019.3
radon==4.0.0
six==1.14.0
text-unidecode==1.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
Werkzeug==0.16.0
